Added coding_report command
The coding_report command (alias "report") gives a brief overview of the coding progress. It currently shows how many abstracts and blocks have been coded by everyone, coder pairs, and individual coders. As the goal is not necessarily to code all abstracts, it does not show a percentage of completed with respect to some goal.